# Optimize Redis cleanup commands