--- a/cmd/authgear/cmd/cmdinternal/internal_redis.go
+++ b/cmd/authgear/cmd/cmdinternal/internal_redis.go
@@ -1,8 +1,11 @@
 package cmdinternal
 
 import (
+	"fmt"
 	"log"
 	"os"
+	"strings"
+	"time"
 
 	"github.com/spf13/cobra"
 
@@ -16,14 +19,11 @@ func init() {
 	cmdInternal.AddCommand(cmdInternalRedis)
 	cmdInternalRedis.AddCommand(cmdInternalRedisCleanUpNonExpiringKeys)
 
-	cmdInternalRedisCleanUpNonExpiringKeys.AddCommand(cmdInternalRedisCleanUpNonExpiringKeysAccessEvents)
-	cmdInternalRedisCleanUpNonExpiringKeys.AddCommand(cmdInternalRedisCleanUpNonExpiringKeysSessionHashes)
-
-	binder.BindString(cmdInternalRedisCleanUpNonExpiringKeysAccessEvents.Flags(), authgearcmd.ArgRedisURL)
-	_ = cmdInternalRedisCleanUpNonExpiringKeysAccessEvents.Flags().Bool("dry-run", true, "Dry-run or not.")
-
-	binder.BindString(cmdInternalRedisCleanUpNonExpiringKeysSessionHashes.Flags(), authgearcmd.ArgRedisURL)
-	_ = cmdInternalRedisCleanUpNonExpiringKeysSessionHashes.Flags().Bool("dry-run", true, "Dry-run or not.")
+	binder.BindString(cmdInternalRedisCleanUpNonExpiringKeys.Flags(), authgearcmd.ArgRedisURL)
+	_ = cmdInternalRedisCleanUpNonExpiringKeys.Flags().Bool("dry-run", true, "Dry-run or not.")
+	_ = cmdInternalRedisCleanUpNonExpiringKeys.Flags().String("scan-count", "100", "Redis SCAN count")
+	_ = cmdInternalRedisCleanUpNonExpiringKeys.Flags().String("expiration", "", "A Go duration literal.")
+	_ = cmdInternalRedisCleanUpNonExpiringKeys.Flags().String("key-pattern", "", fmt.Sprintf("One of %v", strings.Join(cmdredis.KnownKeyPatterns, ",")))
 }
 
 var cmdInternalRedis = &cobra.Command{
@@ -31,32 +31,10 @@ var cmdInternalRedis = &cobra.Command{
 	Short: "Redis commands",
 }
 
-// According to https://linear.app/authgear/issue/DEV-1325/properly-clean-up-redis-keys
-// There are some known key patterns that are non-expiring.
-//
-//	failed-attempts
-//	session-list
-//	access-events
-//	lockout
-//	offline-grant-list
-//
-// For failed-attempts, lockout, the number of them should be small enough to be ignored.
-// They can also be deleted with redis-cli, given we have list-non-expiring-keys.
-//
-//	$ authgear internal redis list-non-expiring-keys >output
-//	$ grep <output 'failed-attempts|lockout' >targets
-//	$ xargs <targets redis-cli del
-//
-// For access-events, it is handled by the subcommand access-events.
-// For session-list and offline-grant-list, it is handled by the subcommand session-hashes.
+// See https://linear.app/authgear/issue/DEV-1325/properly-clean-up-redis-keys
 var cmdInternalRedisCleanUpNonExpiringKeys = &cobra.Command{
 	Use:   "clean-up-non-expiring-keys",
 	Short: "Clean up non-expiring keys.",
-}
-
-var cmdInternalRedisCleanUpNonExpiringKeysAccessEvents = &cobra.Command{
-	Use:   "access-events",
-	Short: "Clean up non-expiring 'app:*:access-events:*'",
 	RunE: func(cmd *cobra.Command, args []string) error {
 		binder := authgearcmd.GetBinder()
 
@@ -70,34 +48,19 @@ var cmdInternalRedisCleanUpNonExpiringKeysAccessEvents = &cobra.Command{
 			return err
 		}
 
+		scanCountString := cmd.Flags().Lookup("scan-count").Value.String()
+		expirationString := cmd.Flags().Lookup("expiration").Value.String()
+		keyPattern := cmd.Flags().Lookup("key-pattern").Value.String()
 		dryRun := cmd.Flags().Lookup("dry-run").Value.String() == "true"
-		err = cmdredis.CleanUpNonExpiringKeysAccessEvents(cmd.Context(), redisClient, dryRun, os.Stdout, log.Default())
-		if err != nil {
-			return err
-		}
+		now := time.Now().UTC()
 
-		return nil
-	},
-}
-
-var cmdInternalRedisCleanUpNonExpiringKeysSessionHashes = &cobra.Command{
-	Use:   "session-hashes",
-	Short: "Clean up non-expiring 'app:*:session-list:*' and 'app:*:offline-grant-list:*'",
-	RunE: func(cmd *cobra.Command, args []string) error {
-		binder := authgearcmd.GetBinder()
-
-		redisURL, err := binder.GetRequiredString(cmd, authgearcmd.ArgRedisURL)
-		if err != nil {
-			return err
-		}
-
-		redisClient, err := cmdredis.NewClient(redisURL)
-		if err != nil {
-			return err
-		}
-
-		dryRun := cmd.Flags().Lookup("dry-run").Value.String() == "true"
-		err = cmdredis.CleanUpNonExpiringKeysSessionHashes(cmd.Context(), redisClient, dryRun, os.Stdout, log.Default())
+		err = cmdredis.CleanUpNonExpiringKeys(cmd.Context(), redisClient, os.Stdout, log.Default(), cmdredis.CleanUpNonExpiringKeysOptions{
+			ScanCountString:  scanCountString,
+			KeyPattern:       keyPattern,
+			ExpirationString: expirationString,
+			DryRun:           dryRun,
+			Now:              now,
+		})
 		if err != nil {
 			return err
 		}