--- a/cmd/authgear/redis/clean_up_non_expiring_keys.go
+++ b/cmd/authgear/redis/clean_up_non_expiring_keys.go
@@ -6,200 +6,178 @@ import (
 	"io"
 	"log"
 	"slices"
-	"strings"
+	"strconv"
+	"time"
 
 	goredis "github.com/redis/go-redis/v9"
 )
 
-func accessEventsToSession(key string) string {
-	return strings.ReplaceAll(key, ":access-events:", ":session:")
+var KnownKeyPatterns = []string{
+	"session-list",
+	"offline-grant-list",
+	"failed-attempts",
+	"access-events",
+	"lockout",
 }
 
-func accessEventsToOfflineGrant(key string) string {
-	return strings.ReplaceAll(key, ":access-events:", ":offline-grant:")
+type CleanUpNonExpiringKeysOptions struct {
+	ScanCountString  string
+	KeyPattern       string
+	ExpirationString string
+	DryRun           bool
+	Now              time.Time
 }
 
-func CleanUpNonExpiringKeysAccessEvents(ctx context.Context, redisClient *goredis.Client, dryRun bool, stdout io.Writer, logger *log.Logger) (err error) {
+func toRedisKeyPattern(keyPattern string) string {
+	return fmt.Sprintf("app:*:%v:*", keyPattern)
+}
+
+func isCandidateForExpire(ctx context.Context, conn *goredis.Conn, key string, isSessionHash bool, now time.Time) (ok bool, err error) {
+	if !isSessionHash {
+		var ttl time.Duration
+		ttl, err = conn.TTL(ctx, key).Result()
+		if err != nil {
+			return
+		}
+		switch ttl {
+		case -2:
+			// -2 means the key does not exist.
+			// Just ignore it.
+			break
+		case -1:
+			// -1 means the key does not expire.
+			ok = true
+		default:
+			// Other values mean the key does have TTL and is going to expire some time in the future.
+			break
+		}
+	} else {
+		var hash map[string]string
+		hash, err = conn.HGetAll(ctx, key).Result()
+		if err != nil {
+			return
+		}
+
+		// The hash has no entries.
+		// So it can expire.
+		if len(hash) == 0 {
+			ok = true
+		} else {
+			hasSessionThatWillExpireInTheFuture := false
+			// The values are time.Time.MarshalText()
+			// We check if all values are timestamp in the past.
+			// If yes, then it can expire.
+			for _, rfc3339 := range hash {
+				var t time.Time
+				err = t.UnmarshalText([]byte(rfc3339))
+				if err != nil {
+					return
+				}
+				if t.After(now) {
+					hasSessionThatWillExpireInTheFuture = true
+				}
+			}
+			if !hasSessionThatWillExpireInTheFuture {
+				ok = true
+			}
+		}
+	}
+	return
+}
+
+func CleanUpNonExpiringKeys(ctx context.Context, redisClient *goredis.Client, stdout io.Writer, logger *log.Logger, options CleanUpNonExpiringKeysOptions) (err error) {
+	// https://redis.io/docs/latest/commands/scan/
+	// The default is 10.
+	// It makes no sense to use a value less than 10.
+	scanCount, err := strconv.ParseInt(options.ScanCountString, 10, 64)
+	if err != nil {
+		err = fmt.Errorf("SCAN count must be an integer: %v", options.ScanCountString)
+		return
+	}
+	if scanCount < 10 {
+		err = fmt.Errorf("SCAN count must greater than or equal to 10: %v", scanCount)
+		return
+	}
+
+	expiration, err := time.ParseDuration(options.ExpirationString)
+	if err != nil {
+		err = fmt.Errorf("expiration must be a valid Go duration literal: %v", options.ExpirationString)
+		return
+	}
+	if expiration < 0 {
+		err = fmt.Errorf("expiration cannot be less than 0: %v", expiration)
+		return
+	}
+
+	if !slices.Contains(KnownKeyPatterns, options.KeyPattern) {
+		err = fmt.Errorf("unsupported key patttern: %v", options.KeyPattern)
+		return
+	}
+
+	if options.Now.IsZero() {
+		err = fmt.Errorf("now cannot be zero")
+		return
+	}
+
+	isSessionHash := false
+	switch options.KeyPattern {
+	case "session-list":
+		isSessionHash = true
+	case "offline-grant-list":
+		isSessionHash = true
+	}
+
 	conn := redisClient.Conn()
 	defer conn.Close()
 
-	// We first scan the key pattern "app:*:access-events:*"
-	// For each key, we see if it has a corresponding "app:*:session:*" or "app:*:offline-grant:*"
-	// If it does, then the session does not expire and its access-events should be kept.
-	// Otherwise, the session has expired, and its access-events can be deleted.
-
 	// We start with a cursor of 0.
 	var cursor uint64
-	pattern := "app:*:access-events:*"
-	// We scan 100 keys
-	var count int64 = 100
+	pattern := toRedisKeyPattern(options.KeyPattern)
 
-	var keysToBeDelete []string
 	var scannedKeyCount = 0
+	var expiredKeyCount = 0
 	for {
 		var keys []string
 		var nextCursor uint64
-		keys, nextCursor, err = conn.Scan(ctx, cursor, pattern, count).Result()
+		keys, nextCursor, err = conn.Scan(ctx, cursor, pattern, scanCount).Result()
 		if err != nil {
 			return
 		}
 
 		scannedKeyCount += len(keys)
-		logger.Printf("SCAN with cursor %v: %v\n", cursor, scannedKeyCount)
+		logger.Printf("SCAN %v %v scanned_total=%v expired_total=%v\n", cursor, pattern, scannedKeyCount, expiredKeyCount)
 
 		for _, key := range keys {
-			offlineGrantKey := accessEventsToOfflineGrant(key)
-			sessionKey := accessEventsToSession(key)
-
-			var existsCount int64
-			existsCount, err = conn.Exists(ctx, offlineGrantKey, sessionKey).Result()
+			var ok bool
+			ok, err = isCandidateForExpire(ctx, conn, key, isSessionHash, options.Now)
 			if err != nil {
 				return
 			}
-
-			// No keys exist. The session is gone. This key can be deleted.
-			if existsCount == 0 {
-				keysToBeDelete = append(keysToBeDelete, key)
+			if ok {
+				expiredKeyCount += 1
+				// -1 means the key does not expire.
+				if options.DryRun {
+					logger.Printf("(dry-run) EXPIRE %v %v\n", key, expiration.Seconds())
+				} else {
+					_, err = conn.Expire(ctx, key, expiration).Result()
+					if err != nil {
+						return
+					}
+					logger.Printf("EXPIRE %v %v\n", key, expiration.Seconds())
+				}
+				fmt.Fprintf(stdout, "%v\n", key)
 			}
 		}
 
 		// According to the doc https://redis.io/docs/latest/commands/scan/
 		// a cursor of 0 means the scan is complete.
 		if nextCursor == 0 {
-			logger.Printf("done SCAN: %v\n", scannedKeyCount)
+			logger.Printf("done scanned_total=%v expired_total=%v\n", scannedKeyCount, expiredKeyCount)
 			break
 		} else {
 			cursor = nextCursor
 		}
-	}
-
-	// Sort the keys.
-	slices.Sort(keysToBeDelete)
-
-	for _, key := range keysToBeDelete {
-		if !dryRun {
-			_, err = conn.Del(ctx, key).Result()
-			if err != nil {
-				return
-			}
-			logger.Printf("deleted %v\n", key)
-		} else {
-			logger.Printf("would delete %v\n", key)
-		}
-		// We prints to stdout anyway.
-		fmt.Fprintf(stdout, "%v\n", key)
 	}
 
 	return nil
-}
-
-func findUpNonExpiringKeysSessionHashesOfKeyPattern(ctx context.Context, conn *goredis.Conn, keyPattern string, stdout io.Writer, logger *log.Logger) (keysToBeReturned []string, err error) {
-	// We first scan the given key pattern.
-	// For each key, we call HGETALL to see if all the key in the hash is NOT a Redis key.
-	// If that is true, this means this hash is OK to delete.
-
-	// We start with a cursor of 0.
-	var cursor uint64
-	// We scan 100 keys
-	var count int64 = 100
-
-	var scannedKeyCount = 0
-	for {
-		var keysContainingHashes []string
-		var nextCursor uint64
-		keysContainingHashes, nextCursor, err = conn.Scan(ctx, cursor, keyPattern, count).Result()
-		if err != nil {
-			return
-		}
-
-		scannedKeyCount += len(keysContainingHashes)
-		logger.Printf("SCAN %v with cursor %v: %v\n", keyPattern, cursor, scannedKeyCount)
-
-		for _, keyToAHash := range keysContainingHashes {
-			var hash map[string]string
-			hash, err = conn.HGetAll(ctx, keyToAHash).Result()
-			if err != nil {
-				return
-			}
-
-			// The hash has no keys. OK to delete.
-			// Actually this will not happen, as Redis says hash will be deleted when
-			// the last field is deleted.
-			if len(hash) == 0 {
-				keysToBeReturned = append(keysToBeReturned, keyToAHash)
-				// Leave this loop early.
-				continue
-			}
-
-			// When we reach here, the hash has at least one key.
-			// We need to prove all keys of the hash point to non-existent session.
-			foundSessionStillExists := false
-
-			// The value is the expiry. We could look at that too.
-			// But looking up the key directly is more straight forward.
-			for keyToAHashField := range hash {
-				var existsCount int64
-				existsCount, err = conn.Exists(ctx, keyToAHashField).Result()
-				if err != nil {
-					return
-				}
-
-				// This session still exists. We can conclude this hash cannot be deleted.
-				if existsCount != 0 {
-					foundSessionStillExists = true
-					break
-				}
-			}
-
-			if !foundSessionStillExists {
-				keysToBeReturned = append(keysToBeReturned, keyToAHash)
-			}
-		}
-
-		// According to the doc https://redis.io/docs/latest/commands/scan/
-		// a cursor of 0 means the scan is complete.
-		if nextCursor == 0 {
-			logger.Printf("done SCAN %v: %v\n", keyPattern, scannedKeyCount)
-			break
-		} else {
-			cursor = nextCursor
-		}
-	}
-
-	return
-}
-
-func CleanUpNonExpiringKeysSessionHashes(ctx context.Context, redisClient *goredis.Client, dryRun bool, stdout io.Writer, logger *log.Logger) (err error) {
-	conn := redisClient.Conn()
-	defer conn.Close()
-
-	sessionListKeys, err := findUpNonExpiringKeysSessionHashesOfKeyPattern(ctx, conn, "app:*:session-list:*", stdout, logger)
-	if err != nil {
-		return
-	}
-
-	offlineGrantListKeys, err := findUpNonExpiringKeysSessionHashesOfKeyPattern(ctx, conn, "app:*:offline-grant-list:*", stdout, logger)
-	if err != nil {
-		return
-	}
-
-	keysToBeDelete := slices.Concat(sessionListKeys, offlineGrantListKeys)
-	// Sort the keys.
-	slices.Sort(keysToBeDelete)
-
-	for _, key := range keysToBeDelete {
-		if !dryRun {
-			_, err = conn.Del(ctx, key).Result()
-			if err != nil {
-				return
-			}
-			logger.Printf("deleted %v\n", key)
-		} else {
-			logger.Printf("would delete %v\n", key)
-		}
-		// We prints to stdout anyway.
-		fmt.Fprintf(stdout, "%v\n", key)
-	}
-
-	return
 }