ref DEV-2443

I have prepared a simple program to generate data.

```
package main

import (
	"context"
	"flag"
	"fmt"
	"time"

	goredis "github.com/redis/go-redis/v9"
)

var t_expired = time.Date(2025, 3, 5, 0, 0, 0, 0, time.UTC)

func doMain() (err error) {
	var redisURL string
	flag.StringVar(&redisURL, "redis-url", "", "URL to redis")

	var count int
	flag.IntVar(&count, "count", 0, "number of data to generate")

	flag.Parse()

	opts, err := goredis.ParseURL(redisURL)
	if err != nil {
		return
	}

	ctx := context.Background()
	redisClient := goredis.NewClient(opts)

	t_expired_bytes, err := t_expired.MarshalText()
	if err != nil {
		return
	}

	for i := range count {
		_, err = redisClient.Set(ctx, fmt.Sprintf("app:test:access-events:%v", i), "not-important", 0).Result()
		if err != nil {
			return
		}

		_, err = redisClient.Set(ctx, fmt.Sprintf("app:test:failed-attempts:purpose:%v", i), "not-important", 0).Result()
		if err != nil {
			return
		}

		_, err = redisClient.Set(ctx, fmt.Sprintf("app:test:lockout:%v", i), "not-important", 0).Result()
		if err != nil {
			return
		}

		_, err = redisClient.HSet(ctx, fmt.Sprintf("app:test:session-list:%v", i), "non-important", t_expired_bytes).Result()
		if err != nil {
			return
		}

		_, err = redisClient.HSet(ctx, fmt.Sprintf("app:test:offline-grant-list:%v", i), "non-important", t_expired_bytes).Result()
		if err != nil {
			return
		}
	}

	return
}

func main() {
	err := doMain()
	if err != nil {
		panic(err)
	}
}

```

I used this program to generate 100,000 * 5 = 500,000 keys

```
go run ./dev2443/main.go -redis-url 'redis://localhost:6379/3' -count 100000
make build TARGET=authgear BIN_NAME=authgear
time sh -c "./authgear internal redis clean-up-non-expiring-keys --key-pattern offline-grant-list --expiration 1s --redis-url 'redis://localhost:6379/3' --dry-run=false 1>/dev/null 2>/dev/null"
time sh -c "./authgear internal redis clean-up-non-expiring-keys --key-pattern session-list --expiration 1s --redis-url 'redis://localhost:6379/3' --dry-run=false 1>/dev/null 2>/dev/null"
time sh -c "./authgear internal redis clean-up-non-expiring-keys --key-pattern lockout --expiration 1s --redis-url 'redis://localhost:6379/3' --dry-run=false 1>/dev/null 2>/dev/null"
time sh -c "./authgear internal redis clean-up-non-expiring-keys --key-pattern access-events --expiration 1s --redis-url 'redis://localhost:6379/3' --dry-run=false 1>/dev/null 2>/dev/null"
time sh -c "./authgear internal redis clean-up-non-expiring-keys --key-pattern failed-attempts --expiration 1s --redis-url 'redis://localhost:6379/3' --dry-run=false 1>/dev/null 2>/dev/null"
```

The output was

```
________________________________________________________
Executed in   33.70 secs    fish           external
   usr time    2.66 secs    0.08 millis    2.66 secs
   sys time    6.76 secs    1.29 millis    6.76 secs


________________________________________________________
Executed in   34.16 secs    fish           external
   usr time    2.70 secs   35.00 micros    2.70 secs
   sys time    6.84 secs  489.00 micros    6.84 secs


________________________________________________________
Executed in   31.72 secs    fish           external
   usr time    2.49 secs   35.00 micros    2.49 secs
   sys time    6.50 secs  470.00 micros    6.50 secs


________________________________________________________
Executed in   33.52 secs    fish           external
   usr time    2.57 secs   40.00 micros    2.57 secs
   sys time    6.76 secs  497.00 micros    6.76 secs


________________________________________________________
Executed in   30.94 secs    fish           external
   usr time    2.43 secs   47.00 micros    2.43 secs
   sys time    6.37 secs  511.00 micros    6.37 secs
```

So it roughly takes 30 seconds to SCAN and EXPIRE 100,000 keys, that is, **3000 keys per second**

I tried change SCAN count from 100 to 1000, the result is similar.